# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: LankaJobs mobile, admin, authentication, and moderation API
servers:
  - url: /api
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: location, in: query, schema: { type: string } }
        - { name: employmentType, in: query, schema: { type: string } }
        - { name: workMode, in: query, schema: { type: string } }
      responses:
        "200": { description: Paginated published jobs }
    post:
      operationId: createJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "201": { description: Draft job }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /jobs/mine:
    get:
      operationId: listMyJobs
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current user's jobs }
  /jobs/recommendations:
    get:
      operationId: listRecommendations
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Rule-based profile recommendations }
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Published job }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: updateJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobInput" }
      responses:
        "200": { description: Updated owned job }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted owned job }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{id}/submit:
    post:
      operationId: submitJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Submitted for moderation }
  /saved-jobs:
    get:
      operationId: listSavedJobs
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current user's saved jobs }
  /saved-jobs/{jobId}:
    post:
      operationId: saveJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Saved }
    delete:
      operationId: unsaveJob
      tags: [jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
...
    Unauthorized: { description: Authentication required }
...
    NotFound: { description: Resource not found }
...
  schemas:
...
    JobInput:
      type: object
      required: [title, company, description, location, employmentType, workMode]
      properties:
        title: { type: string }
        company: { type: string }
        description: { type: string }
        categoryId: { type: string }
        location: { type: string }
        employmentType: { type: string }
        workMode: { type: string }
        salary: { type: string }
        experience: { type: string }
        education: { type: string }
        skills: { type: array, items: { type: string } }
        applicationEmail: { type: string, format: email }
        applicationUrl: { type: string, format: uri }
        closingDate: { type: string, format: date }